[css-grid] Avoid the estimated grid area on intrinsic size computation
We perform a pre-layout, based on a estimation of the grid area, of the
grid items in two different situations: orthogonal and baseline aligned
items.

We do this pre-layout operation during both, intrinsic size and layout.
Since this pre-layout operation may alter the size of the grid item, we
must ensure the pre-layout done during intrinsic size computation is
always followed by a full layout of the grid container.

This CL adds information about the phase of the layout and avoid using
the estimated grid area if the intrinsic width operation is not asked
as part of the grid container's layout.

Bug: 1086132
Change-Id: I73c39e6c3ad6cd74aa50fe33106e25cd63b7625f