enable wpt tests with mock screen enumeration
This CL enables possibility to isolate blink implementation of
screen enumeration from the rest of the stack.

It provides the possibility to create mock displays at test conveniences
by rerouting the GetDisplay call to the JS implementation in the test.

Existing test has been modified to use this test framework.

Bug: 1075656
Change-Id: I1a38426deaccff14036f00806ef7dfc878bcd4ce